Add Swipe Edit / Drag and Drop

+++ app/src/main/java/com/espfullstack/wedoo/helper/SwipeCallback.java
@@ -19,14 +19,16 @@ public class SwipeCallback extends ItemTouchHelper.SimpleCallback {
 
     private ToDooAdapter toDooAdapter;
 
-    private Drawable deleteIcon;
-    private final ColorDrawable deleteBackground;
+    private Drawable deleteIcon, editIcon;
+    private final ColorDrawable deleteBackground, editBackground;
 
     public SwipeCallback(ToDooAdapter toDooAdapter, Context context) {
-        super(0, ItemTouchHelper.LEFT);
+        super(0, ItemTouchHelper.LEFT | ItemTouchHelper.RIGHT);
         this.toDooAdapter = toDooAdapter;
         deleteIcon = context.getDrawable(R.drawable.ic_trash);
         deleteBackground = new ColorDrawable(Color.RED);
+        editIcon = context.getDrawable(R.drawable.ic_edit);
+        editBackground = new ColorDrawable(context.getColor(R.color.amber));
     }
 
     @Override
@@ -40,7 +42,29 @@ public class SwipeCallback extends ItemTouchHelper.SimpleCallback {
     @Override
     public void onSwiped(@NonNull RecyclerView.ViewHolder viewHolder, int direction) {
         int position = viewHolder.getAdapterPosition();
-        toDooAdapter.deleteItem(position);
+        if(direction == ItemTouchHelper.LEFT) {
+            toDooAdapter.deleteItem(position);
+        } else if(direction == ItemTouchHelper.RIGHT){
+            toDooAdapter.editItem(position);
+        }
+    }
+
+
+    @Override
+    public boolean isLongPressDragEnabled() {
+        return true;
+    }
+
+    @Override
+    public boolean isItemViewSwipeEnabled() {
+        return true;
+    }
+
+    @Override
+    public int getMovementFlags(@NonNull RecyclerView recyclerView, @NonNull RecyclerView.ViewHolder viewHolder) {
+        int dragFlags = ItemTouchHelper.UP | ItemTouchHelper.DOWN;
+        int swipeFlags = ItemTouchHelper.LEFT | ItemTouchHelper.RIGHT;
+        return makeMovementFlags(dragFlags, swipeFlags);
     }
 
     @Override
@@ -50,19 +74,38 @@ public class SwipeCallback extends ItemTouchHelper.SimpleCallback {
         if (viewHolder.getAdapterPosition() == -1) {
             return;
         }
+
         float margin = AndroidUtils.convertDpToPixel(8, recyclerView.getContext());
 
-        deleteBackground.setBounds(itemView.getRight() + ((int) dX), itemView.getTop() + (int)margin, itemView.getRight(), itemView.getBottom());
-        deleteBackground.draw(c);
+        ColorDrawable background = deleteBackground;
+        Drawable icon = deleteIcon;
 
-        int iconMargin = (itemView.getHeight() - deleteIcon.getIntrinsicHeight()) / 2;
-        int iconTop = itemView.getTop() + (itemView.getHeight() - deleteIcon.getIntrinsicHeight()) / 2;
-        int iconBottom = iconTop + deleteIcon.getIntrinsicHeight();
+        int iconMargin = (itemView.getHeight() - icon.getIntrinsicHeight()) / 2;
+        int iconTop = itemView.getTop() + (itemView.getHeight() - icon.getIntrinsicHeight()) / 2;
+        int iconBottom = iconTop + icon.getIntrinsicHeight();
 
-        int iconLeft = itemView.getRight() - iconMargin - deleteIcon.getIntrinsicWidth();
-        int iconRight = itemView.getRight() - iconMargin;
-        deleteIcon.setBounds(iconLeft, iconTop, iconRight, iconBottom);
-        deleteIcon.draw(c);
+        if(dX > 0) {
+            background = editBackground;
+            icon = editIcon;
+            int iconLeft = itemView.getLeft() + iconMargin;
+            int iconRight = iconLeft + icon.getIntrinsicWidth();
+            background.setBounds(itemView.getLeft(), itemView.getTop() + (int) margin, itemView.getLeft() + ((int) dX), itemView.getBottom());
+            icon.setBounds(iconLeft, iconTop, iconRight, iconBottom);
+
+        } else if (dX < 0){
+            background = deleteBackground;
+            icon = deleteIcon;
+            int iconLeft = itemView.getRight() - iconMargin - icon.getIntrinsicWidth();
+            int iconRight = itemView.getRight() - iconMargin;
+            background.setBounds(itemView.getRight() + ((int) dX), itemView.getTop() + (int) margin, itemView.getRight(), itemView.getBottom());
+            icon.setBounds(iconLeft, iconTop, iconRight, iconBottom);
+        } else {
+            background.setBounds(0, 0,0,0);
+        }
+
+        background.draw(c);
+        icon.draw(c);
+
         super.onChildDraw(c, recyclerView, viewHolder, dX, dY, actionState, isCurrentlyActive);
     }
 }
